Better rails runner output for missing files
This updates The Rails runner command to give a more helpful error
message when given a path as input but the file does not exist.

Instead of the cryptic "undefined local variable or method" the user is
told that the file could not be found.

I've gone for the simplest approach I could think of to identify when
the input is likely a file path: it ends with `".rb"`. This isn't a
perfect approach – there's nothing stopping someone passing a script
without an extension, and it's also possible to have valid Ruby syntax
that ends in `".rb" – but it does feel "good enough" and will make the
Rails runner slightly more user-friendly for those working with file
paths without negatively impacting those that are passing in Ruby
code.