Use a static list of usable ruby versions
Instead of a Docker API call.

CI was down because that docker API is being deprecated.
I started upgrading to the V2 API, but it only returns 100
tags per page, and there is currently almost 1k tags.

That doesn't seem sustainable, so I think a simple hardcoded
list is more than good enough.