Pass options accessor to Cache#fetch block
Closes: https://github.com/rails/rails/pull/45080

This is a surprisingly common use case when dealing with APIs
with short lived authentication tokens.

This implementation is very much the same as the original one
by Andrii, except that we yield an object with a restricted interface
instead of yielding the options hash.

The reason is that not all options can be changed between the read
and the write, so by restricting the interface we avoid surprising
users.

Co-Authored-By: Andrii Gladkyi <arg@arg.zone>