Make ActionView::PathSet immutable
Previously we would mutate the PathSet in order to implement
{append,prepend}_view_path.

This removes the mutation methods we had and instead constructs a new
PathSet whenever we are modifying it. This should allow flexibility in
the classes holding a view_path to know that their view paths isn't
modified (for example, to allow caching).