Add regression test for broken filter
In 7690290, the filter was not finding the end of the methods defined
using the `test do` syntax. This would cause more than 1 test to be
selected depending on the line number passed as argument.

The previous test wasn't failing because the two tests asked to run
where one after the other and the filter, even broken would only
run those two tests. By changing the test definition order we trigger
the bug introduced by that commit.