Allow empty callstack in AS::Deprecation#warn
This fixes an "undefined method 'path' for nil:NilClass" error in
`Deprecation::Reporting#extract_callstack` when the callstack is empty,
either because it was specified as such or because the `warn` call was
not sufficiently deep in a call tree (e.g. at the top level of a new
thread).