Minor refactoring to schema migration and internal metadata
This clean up wraps the messages at 80 characters and clarifies the
changes needed. 99% of apps won't see this warning as they aren't
accessing MigrationContext directly. This is only for libraries or apps
that are passing a custom `SchemaMigration` or `InternalMetadata` class
so they get a more helpful error.

In addition I moved the `NullSchemaMigration` up to the top of the class
to match InternalMetadata.