Pass deprecator to AS::Deprecation callbacks
This commit adds support for 3-arity `ActiveSupport::Deprecation`
behavior callbacks, which receive the deprecator instance as their third
argument.  This allows such callbacks to change their behavior based on
the deprecator's state.  For example, based on the deprecator's `debug`
flag.

This commit also fixes the `:stderr` and `:log` behaviors to check
the given deprecator's `debug` flag, instead of the global
`ActiveSupport::Deprecation.debug` flag.

Note that 2-arity and 4-arity callbacks will continue to behave the
same as before.  However, callbacks with -1 and -2 arity (i.e. with
splat args) will now receive the deprecator as their third argument,
instead of the deprecation horizon and gem name.