fix some issues in failing CI
* fixes most of the db tasks tests
* fixes missing Ar Base to pass into `migrate`

But this points out that we are probably never falling through to
Base.connection since TemporaryConnection could exist. This means that
we might not be able to use the fallback, if temp connection exists _at
all_ we'll pick it up. So we either need to always be dropping it and
re-establishing (gross) or actually set a "current connection" somehow
and raise if they are different.

We might have a problem here for apps using Base.connection that super
into the Rails tasks which won't have Base.connection. Maybe we can
assign an existing connection (like have a MigrationConnection and a
Base and then have a with_temporary_connection that holds reference?)