Use `sslmode` option instead of `ssl_mode` for MySQL command connections
In f36c261a44 support for the ssl-mode option was added for MySQL when
using the dbconsole command and MySQLDatabaseTasks.
However, the mysql2 gem uses `ssl_mode` instead of `sslmode`:

https://github.com/brianmario/mysql2/blob/ba4d46551d132492b34205cdb8fa224c92765bef/lib/mysql2/client.rb#L51

As `ssl_mode` is what should be used in the database.yml configuration
for MySQL we should use it as well for the dbconsole command and
MySQLDatabaseTasks.

Co-authored-by: Chris Gunther <chris@room118solutions.com>