Fix plugin generation inside applications
Previously, generating a plugin inside an application would throw an
error when trying to create a dummy_app due to the generator trying to
reuse the application's name.

This is fixed by extracting the logic used to determine the source of an
application's name, and then only using the existing name as the source
for the app:update task.