Keep rescue template paths in an array
Gotta be honest, this is so I can make some hacks.  Basically I would
like an engine to specify where to find rescue templates, and currently
there's no way to add search paths to the debug view lookup context.
This commit turns the template path in to an array (that I plan to
mutate, but nobody should do that besides me until we make an actual
good API).

I added the `dup` in `initialize` so in case the array is accidentally
mutated we don't leak memory.