convert prepare_all
this method cant use a block version because `create` will re-estblish
the connection and it will get dropped so `pool` will be a dead conn.

we also cant call `pool.connection` when a db doesn't exist so i need to
est the connection and then pick it up later from the class.