Add custom key sizes for db:encryption:init task
This change allows to pass arguments to the
`bin/rails db:encryption:init` task. That way, we ca easily
specify the bytesize of each key. For example, if we want the
`primary_key` and the `deterministic_key` to be 64 bytes; and
the `key_derivation_salt` to be 32 bytes, we can run the
following command:

    bin/rails "db:encryption:init[64,64,32]"

We will then get the following result in the console:

Add this entry to the credentials of the target environment:

active_record_encryption:
  primary_key: kkBkJb6LBYCgITxu80LW4KU3c8Pm4zSF7CknrE836pGsBVMrtAoENNFya3pDzASK
  deterministic_key: rFOHbNpRsvaSqYSuhyRGiWFKPMzjb9wdqfRLQZnCkheVcQjidVR7bn6diUJIjtls
  key_derivation_salt: TV2LerVWCeShYxSFjjzAs9pJzx77GNJp