Start hiding the exception object from the views
We are currently mutating exception objects and I would like to stop
doing that.  Unfortunately the views are calling many methods directly
on the exception and expecting that the mutations exist.

This patch refactors the templates so that they ask the ExceptionWrapper
class for information about the exception rather than directly asking
the exception object itself