Remove syntax error exception mutations
I am trying to clean up error handling somewhat and this code is getting
in the way.  It mutates the backtrace on exceptions, but only when the
exceptions were SyntaxError instances.  I believe this codepath isn't
used in real life, so I want to remove it.

In real life, the template compiler will catch `SyntaxError` instances
and re-raise them as `SyntaxErrorInTemplate` instances:

  https://github.com/rails/rails/blob/08ef43ed5e01b42b2aec847a0424515f537ed427/actionview/lib/action_view/template.rb#L349-L360

If you have a syntax error in your controller, it's simply not caught by
the DebugExceptions middleware, and you get the "We're sorry, but
something went wrong" page.

I think we can do something good in the case of syntax errors inside of
controllers, but this code isn't doing it.