Add ActiveRecord.schema_dump_formats
This optional setting allows multiple formats to be dumped if desired. For
example:

```
config.active_record.schema_dump_formats = [:ruby, :sql]
```

Our database contains triggers and functions which cannot be expressed in
`db/schema.rb` so we have to use `db/structure.sql`. But then some of us
appreciate the readability of `db/schema.rb`. So it'd be ideal for us if we can
dump both files in our project.