Add multi-db support for db:version tasks
The `rails db:version` task makes no effort today to display information for multi-database applications.

This brings up this simplistic tasks up to parity with the rest of the DB tasks that support multi-database applications.

This is accomplished by following the pattern of making the non-namespaced task (i.e. `db:version`) to output a multi-line string, iterating through all defined databases for the current environment in `database.yml`

It also adds a namespaced instance of the task for each database defined (e.g. `db:version:animals`).

**A hidden change, but critical (for the purposes that I'm working on), that is happening here** is that database configurations with `database_tasks: false` will be excluded from the logic in this task – due to the usage of the `ActiveRecord::Base.configurations.configs_for` method, which by default has `include_hidden: false`!