Add `ActiveRecord::Base.connection_pool.without_connection(&block)`
In a typical Rails app, we're making more database calls compared to
network calls. It is impractical to wrap every database call with
`with_connection`.

It is more manageable to assume the database connection by default and
have a way to explicitly mark a block of code that doesn't need a
database connection.

Added a new method `without_connection` to `ActiveRecord::ConnectionAdapters::ConnectionPool`
to achieve that.

```ruby
Thread.new do
  Rails.application.executor.wrap do
    user = User.first

    ActiveRecord::Base.connection_pool.without_connection do
      @response = Net::HTTP.get(URI("https://rubyonrails.org/?user_id=#{@user.id}"))
    end

    user.update(description: @response)
  end
end
```