fix: Ensure `ActiveJob::TestHelper` sets `scheduled_at` to a `Float`
I noticed a difference in behavior when inspecting `ActiveJob::Core#scheduled_at` in non-test environments vs when inspecting in while executing tests.

Prior to this change, `ActiveJob::TestHelper` would set `ActiveJob::Core#scheduled_at` to a `Time` object in tests,
however this attribute was set to a `Float` objects in non-test environments:

<https://github.com/rails/rails/blob/d1aa6af5cc093164043641e99e30a9c8c60aa159/activejob/lib/active_job/core.rb#L161>

This change ensures parity in all runtime environments.