MemoryStore: preserve entry ttl when incrementing
Fix: https://github.com/rails/rails/issues/46301

When incrementing in Memcached or Redis, we use `raw` values, so
only the backing store TTL counts, and is preserved.

So for consistency we need to emulate this in MemoryStore by
re-using the existing entry's `expires_at`.

---

Some of the changes introduced in [the original
PR](https://github.com/rails/rails/pull/46305) were discarded, such as
the `normalize_` methods, as well as the change that originated in [that
PR](https://github.com/rails/rails/pull/45068) that sets a value if none
exists.